requirements.txt
Pillow==9.3.0
requests==2.27.1
streamlit==1.16.0
streamlit_lottie==0.0.3
